water prefabs added, basic water placement and river.
Allot of these assets that exist far away will most likely be rendered as a skybox and re-imported to save on file size. The sand dunes and river are very large objects that will be hard to fill in properly without a decent amount of assets. specifically the riverbed of the oasis as it will need plants, trees, and some rocks along the river.